refactor: cli commands
