+++ binaries/cli/src/attach.rs
@@ -15,7 +15,7 @@ use std::{path::PathBuf, sync::mpsc, time::Duration};
 use tracing::{error, info};
 use uuid::Uuid;
 
-use crate::handle_dataflow_result;
+use crate::common::handle_dataflow_result;
 
 pub fn attach_dataflow(
     dataflow: Descriptor,
