+++ binaries/cli/src/commands/graph/.gitignore
@@ -1,0 +1,1 @@
+!*template.html
