+++ binaries/cli/src/commands/graph/mermaid-template.html
