+++ binaries/cli/src/commands/stop.rs
@@ -1,0 +1,111 @@
+use super::{default_tracing, Executable};
+use crate::common::{connect_to_coordinator, handle_dataflow_result, query_running_dataflows};
+use communication_layer_request_reply::TcpRequestReplyConnection;
+use dora_core::topics::{DORA_COORDINATOR_PORT_CONTROL_DEFAULT, LOCALHOST};
+use dora_message::cli_to_coordinator::ControlRequest;
+use dora_message::coordinator_to_cli::ControlRequestReply;
+use duration_str::parse;
+use eyre::{bail, Context};
+use std::net::IpAddr;
+use std::time::Duration;
+use uuid::Uuid;
+
+#[derive(Debug, clap::Args)]
+/// Stop the given dataflow UUID. If no id is provided, you will be able to choose between the running dataflows.
+pub struct Stop {
+    /// UUID of the dataflow that should be stopped
+    uuid: Option<Uuid>,
+    /// Name of the dataflow that should be stopped
+    #[clap(long)]
+    name: Option<String>,
+    /// Kill the dataflow if it doesn't stop after the given duration
+    #[clap(long, value_name = "DURATION")]
+    #[arg(value_parser = parse)]
+    grace_duration: Option<Duration>,
+    /// Address of the dora coordinator
+    #[clap(long, value_name = "IP", default_value_t = LOCALHOST)]
+    coordinator_addr: IpAddr,
+    /// Port number of the coordinator control server
+    #[clap(long, value_name = "PORT", default_value_t = DORA_COORDINATOR_PORT_CONTROL_DEFAULT)]
+    coordinator_port: u16,
+}
+
+impl Executable for Stop {
+    fn execute(self) -> eyre::Result<()> {
+        default_tracing()?;
+        let mut session =
+            connect_to_coordinator((self.coordinator_addr, self.coordinator_port).into())
+                .wrap_err("could not connect to dora coordinator")?;
+        match (self.uuid, self.name) {
+            (Some(uuid), _) => stop_dataflow(uuid, self.grace_duration, &mut *session),
+            (None, Some(name)) => stop_dataflow_by_name(name, self.grace_duration, &mut *session),
+            (None, None) => stop_dataflow_interactive(self.grace_duration, &mut *session),
+        }
+    }
+}
+
+fn stop_dataflow_interactive(
+    grace_duration: Option<Duration>,
+    session: &mut TcpRequestReplyConnection,
+) -> eyre::Result<()> {
+    let list = query_running_dataflows(session).wrap_err("failed to query running dataflows")?;
+    let active = list.get_active();
+    if active.is_empty() {
+        eprintln!("No dataflows are running");
+    } else {
+        let selection = inquire::Select::new("Choose dataflow to stop:", active).prompt()?;
+        stop_dataflow(selection.uuid, grace_duration, session)?;
+    }
+
+    Ok(())
+}
+
+fn stop_dataflow(
+    uuid: Uuid,
+    grace_duration: Option<Duration>,
+    session: &mut TcpRequestReplyConnection,
+) -> Result<(), eyre::ErrReport> {
+    let reply_raw = session
+        .request(
+            &serde_json::to_vec(&ControlRequest::Stop {
+                dataflow_uuid: uuid,
+                grace_duration,
+            })
+            .unwrap(),
+        )
+        .wrap_err("failed to send dataflow stop message")?;
+    let result: ControlRequestReply =
+        serde_json::from_slice(&reply_raw).wrap_err("failed to parse reply")?;
+    match result {
+        ControlRequestReply::DataflowStopped { uuid, result } => {
+            handle_dataflow_result(result, Some(uuid))
+        }
+        ControlRequestReply::Error(err) => bail!("{err}"),
+        other => bail!("unexpected stop dataflow reply: {other:?}"),
+    }
+}
+
+fn stop_dataflow_by_name(
+    name: String,
+    grace_duration: Option<Duration>,
+    session: &mut TcpRequestReplyConnection,
+) -> Result<(), eyre::ErrReport> {
+    let reply_raw = session
+        .request(
+            &serde_json::to_vec(&ControlRequest::StopByName {
+                name,
+                grace_duration,
+            })
+            .unwrap(),
+        )
+        .wrap_err("failed to send dataflow stop_by_name message")?;
+    let result: ControlRequestReply =
+        serde_json::from_slice(&reply_raw).wrap_err("failed to parse reply")?;
+    match result {
+        ControlRequestReply::DataflowStopped { uuid, result } => {
+            handle_dataflow_result(result, Some(uuid))
+        }
+        ControlRequestReply::Error(err) => bail!("{err}"),
+        other => bail!("unexpected stop dataflow reply: {other:?}"),
+    }
+}
